feat(teampage): created Team page without Layout(navbar and footer)
