+++ src/components/TeamMember.jsx
@@ -8,10 +8,10 @@ const TeamMember = ({ title, job, photo }) => {
                 <Image src={photo} alt='My Image' width={176} height={190} />
                 <div className='bg-black w-full h-[5px] mt-0 bg-phosphorescent '></div>
                 <div className='p-5'>
-                    <h5 className='mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white pb-[33px] h-[50px] mb-[40px]'>
+                    <h5 className='mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white pb-[33px] h-[50px] mb-[40px] text-center'>
                         {title}
                     </h5>
-                    <p className='mb-3 font-normal text-gray-700 dark:text-gray-400 pb-[40px]'>
+                    <p className='mb-3 font-normal text-gray-700 dark:text-gray-400 pb-[40px] text-center'>
                         {job}.
                     </p>
                 </div>
