--- src/components/__test__/SingleTeam.test.js
@@ -1,7 +1,0 @@
-import renderer from "react-test-renderer";
-import TeamMember from "@/components/TeamMember";
-
-it("renders correctly", () => {
-    const tree = renderer.create(<TeamMember />).toJSON();
-    expect(tree).toMatchSnapshot();
-});
